refactor: move PriceChart component into charts folder
