+++ src/components/Tokens/TokenDetails/ChartSection.tsx
@@ -7,7 +7,7 @@ import { useAtomValue } from 'jotai/utils'
 import { pageTimePeriodAtom } from 'pages/TokenDetails'
 import { startTransition, Suspense, useMemo } from 'react'
 
-import { PriceChart } from './PriceChart'
+import { PriceChart } from '../../Charts/PriceChart'
 import TimePeriodSelector from './TimeSelector'
 
 function usePriceHistory(tokenPriceData: TokenPriceQuery): PricePoint[] | undefined {
